Authentication updated

+++ backend/base/api/views.py
@@ -1,8 +1,9 @@
 from django.http import JsonResponse
 from rest_framework import permissions
 from rest_framework.response import Response
-from rest_framework.decorators import api_view, permission_classes
+from rest_framework.decorators import api_view, permission_classes, authentication_classes
 from rest_framework.permissions import IsAuthenticated
+from rest_framework.authentication import SessionAuthentication, TokenAuthentication
 
 from rest_framework_simplejwt.serializers import TokenObtainPairSerializer
 from rest_framework_simplejwt.views import TokenObtainPairView
@@ -42,12 +43,13 @@ def getRoutes(request):
 
 
 @api_view(['GET'])
-def getBooks(request, id):
-    user = User.objects.get(id=id)
+@permission_classes([IsAuthenticated])
+def getBooks(request):
+    user = request.user
     if user.role == "Author":
         books = user.book_set.all()
     else:
-        return Response(status=400, data={"Message": "Can't access"})
+        return Response(status=401, data={"Message": "Can't access"})
     serializer = BookSerializer(books, many=True)
     return Response(serializer.data)
 
@@ -67,8 +69,9 @@ def getBook(request, id):
 
 
 @api_view(['POST'])
-def createBook(request, id):
-    user = User.objects.get(id=id)  # needs to be changed
+@permission_classes([IsAuthenticated])
+def createBook(request):
+    user = request.user  # needs to be changed
     book = Book.objects.create(
         title=request.data['title'], description=request.data['description'], author=user, authorName=user.name)
     serializer = BookSerializer(book)
@@ -76,10 +79,14 @@ def createBook(request, id):
 
 
 @api_view(['PUT'])
+@permission_classes([IsAuthenticated])
 def editBook(request, id):
-    # user = User.objects.get(id=1) #needs to be changed
+    user = request.user
     book = Book.objects.get(
         id=id)
+    if book.author != user:
+        return Response(status=401, data={"Message": "Can't access"})
+
     book.title = request.data['title']
     book.description = request.data['description']
     book.save()
@@ -88,18 +95,26 @@ def editBook(request, id):
 
 
 @api_view(['DELETE'])
+@permission_classes([IsAuthenticated])
 def deleteBook(request, id):
     # user = User.objects.get(id=1) #needs to be changed
+    user = request.user
     book = Book.objects.get(
         id=id)
+    if book.author != user:
+        return Response(status=401, data={"Message": "Can't delete"})
     book.delete()
     serializer = BookSerializer(book)
     return Response(serializer.data)
 
 
 @api_view(['POST'])
+@permission_classes([IsAuthenticated])
 def createPage(request, id):
+    user = request.user
     book = Book.objects.get(id=id)
+    if book.author != user:
+        return Response(status=401, data={"Message": "Can't access"})
     page = Page.objects.create(book=book, text=request.data['text'])
     serializer = PageSerializer(page)
     return Response(serializer.data)
@@ -114,10 +129,14 @@ def getAllPages(request, id):
 
 
 @api_view(['PUT'])
+@permission_classes([IsAuthenticated])
 def editPage(request, id):
-    # user = User.objects.get(id=1) #needs to be changed
+    user = request.user
     page = Page.objects.get(
         id=id)
+    book = page.book
+    if book.author != user:
+        return Response(status=401, data={"Message": "Can't access"})
     page.text = request.data['text']
     page.save()
     serializer = PageSerializer(page)
@@ -125,10 +144,14 @@ def editPage(request, id):
 
 
 @api_view(['DELETE'])
+@permission_classes([IsAuthenticated])
 def deletePage(request, id):
-    # user = User.objects.get(id=1) #needs to be changed
+    user = request.user
     page = Page.objects.get(
         id=id)
+    book = page.book
+    if book.author != user:
+        return Response(status=401, data={"Message": "Can't delete"})
     page.delete()
     serializer = PageSerializer(page)
     return Response(serializer.data)
@@ -145,3 +168,10 @@ def createUser(request):
         return Response(serializer.data)
     else:
         return Response("Username already exists", 401)
+
+
+@api_view(['GET'])
+@permission_classes([IsAuthenticated])
+def testToken(request):
+    print(request.user.id)
+    return Response("passed")
